average update
A new serverlist slash command has been added, which outputs brief information about the servers on which the bot is located and their owners, and saves it in JSON. The structure of the messageCreate event has been rewritten. Now, to simplify the addition of interactions, a class has been added, as well as a function that sends a gif to the channel. Added random interaction.